Split content tab in user settings
