+++ frontend/app/pages/user-settings.vue
@@ -6,6 +6,8 @@
 			<m-nav-link to="account" :title="$t('account_tab')"/>
 			<m-nav-link v-if="isMe" to="settings" :title="$t('settings')"/>
 			<m-nav-link v-if="isMe" to="content" :title="$t('content_tab')"/>
+			<m-nav-link v-if="isMe" to="following" :title="$t('following')"/>
+			<m-nav-link v-if="isMe" to="blocking" :title="$t('blocking')"/>
 			<m-nav-link v-if="isMe" to="accounts" :title="$t('connected_accounts_tab')"/>
 			<!-- <m-nav-link to="api" :title="$t('api_access_tab')"/> -->
 			<template #content>
